updated tab menu and pagination components
